SQUASH BEFORE MERGING
This commit will be squashed before merging. It copies code verbatim
from join_funcs.go to lookupjoin/constraint_builder.go to make it easy
to review the changes made when moving code in the following commit.
